move router methods
